Add files via upload
Modyfikacja kodu

+++ Acceptor/src/main/java/com/simplemethod/sonb/Acceptor/model/Vote.java
@@ -5,11 +5,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class Vote {
-    private int sequenceId;
+    private BigInteger sequenceId;
     private String message;
 }
